Setup test suite with bedrock-test@4.
