+++ tests/test.config.js
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2019 Digital Bazaar, Inc. All rights reserved.
+ */
+'use strict';
+
+const {config} = require('bedrock');
+const path = require('path');
+
+config.mocha.tests.push(path.join(__dirname, 'mocha'));
